# API-Spec/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PaymentGateway API",
    "version": "v1"
  },
  "paths": {
    "/Payments/{paymentId}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "name": "PaymentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment was found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPaymentResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/Payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPaymentRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostPaymentRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostPaymentRequestModel"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The payment was created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PostPaymentResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPaymentResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPaymentResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "The request data is invalid",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreditCardModel": {
        "type": "object",
        "properties": {
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "expiryDateMonth": {
            "type": "integer",
            "format": "int32"
          },
          "expiryDateYear": {
            "type": "integer",
            "format": "int32"
          },
          "ccv": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PaymentModel": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "string",
            "format": "uuid"
          },
          "merchantId": {
            "type": "string",
            "format": "uuid"
          },
          "externalShopperIdentifier": {
            "type": "string",
            "nullable": true
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "creditCard": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreditCardModel"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetPaymentResponseModel": {
        "type": "object",
        "properties": {
          "payment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentModel"
              }
            ],
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponseModel": {
        "type": "object",
        "properties": {
          "errorDescription": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PostPaymentRequestModel": {
        "type": "object",
        "additionalProperties": false
      },
      "PostPaymentResponseModel": {
        "type": "object",
        "properties": {
          "wasPaymentSuccessful": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}